% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregations.obeu.R
\name{aggregations.obeu}
\alias{aggregations.obeu}
\title{Aggregate an OBEU dataset}
\usage{
aggregationsobeu(data,what="...", to_what= "variable", fun.aggregate="sum", ...)
}
\arguments{
\item{data}{The input dataset to aggregate}

\item{what}{A string of dimensions to aggregate (e.g."dim1,dim2" that results in dim1+dim2)}

\item{to_what}{A string of dimensions to aggregate (e.g."dim1,dim2" that results in dim1+dim2)}

\item{fun.aggregate}{The aggregation function (e.g.sum,mean,sd,var,max,min,...,summary,decribe).}

\item{...}{Other arguments to pass, see \code{\link{cast}} from \pkg{reshape} package}
}
\value{
A json string with the results of the selected aggregation parameters.
Aggregate \code{what} \code{to_what}.
}
\description{
Aggregate an OBEU dataset according to its dimensions.
}
\details{
Works perfect for every dataset but needs a few slight changes.
}
\examples{
#OBeu Example
aggregations.obeu(Budget_Thessaloniki_2015_Revenue)
aggregations.obeu(Budget_Thessaloniki_2015_Revenue,"Υπηρεσία","variable",sum)
aggregations.obeu(Budget_Thessaloniki_2015_Revenue,"Υπηρεσία","variable",mean)
}
\author{
Kleanthis Koupidis
}
\references{
add
}
\seealso{
add
}

