% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Athens_executed_ts.R
\docType{data}
\name{Athens_executed_ts}
\alias{Athens_executed_ts}
\title{Time series of Executed Expenditure Budget Phase of Municipality of Athens}
\format{A ts object with 12 draft amounts from 2004-2015}
\source{
add #url#
}
\usage{
Athens_executed_ts
}
\description{
\describe
Time series data with the Executed Budget phase expenditure amounts of Municipality of Athens from 2004-2015
\itemize{
  \item {The years of the recorded executed budget phase amounts.}
  \item {The executed budget phase amounts of this time range.}
  }
}
\keyword{datasets}

