% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Athens_reserved_ts.R
\docType{data}
\name{Athens_reserved_ts}
\alias{Athens_reserved_ts}
\title{Time series of Reserved Expenditure Budget Phase of Municipality of Athens}
\format{A ts object with 12 reserved amounts from 2004-2015}
\source{
add #url#
}
\usage{
Athens_reserved_ts
}
\description{
\describe
Time series data with the Reserved Budget phase expenditure amounts of Municipality of Athens from 2004-2015
\itemize{
  \item {The years of the recorded reserved budget phase amounts.}
  \item {The reserved budget phase amounts of this time range.}
  }
}
\keyword{datasets}

