% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babbage.ts.obeu.R
\name{babbage.ts.obeu}
\alias{babbage.ts.obeu}
\title{Reading and analyze babbage time series data}
\usage{
babbage.ts.obeu(json_data,time,amount,prediction_step)
}
\arguments{
\item{json_data}{The json output from babbage api.}

\item{time}{Specify the time label of the json time series data.}

\item{amount}{Specify the amount label of the json time series data.}

\item{prediction_steps}{The number of prediction steps}
}
\value{
A json string with the resulted parameters of the ts.obeu.
}
\description{
EXtract and analyze the time series data from babbage api, using the ts.obeu function.
}
\details{
This function extracts the time series data from a json file resulted from the babbage api
and then analyze it using the ts.obeu function.
}
\examples{
#Not YET an OBeu Example


}
\author{
Kleanthis Koupidis
}
\references{
add
}
\seealso{
add
}

