% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Budget_Thessaloniki_2015_Revenue.R
\docType{data}
\name{Budget_Thessaloniki_2015_Revenue}
\alias{Budget_Thessaloniki_2015_Revenue}
\title{Revenue Details of Municipality of Thessaloniki for 2015}
\format{A data frame with 53940 rows and 10 variables.}
\source{
\url{ ... }
}
\usage{
Budget_Thessaloniki_2015_Revenue
}
\description{
\describe
A dataset containing the Budget phase revenue amounts of Municipality of Thessaloniki
}
\details{
\itemize{
  \item{administrative unit}{No admin unit in revenue}
  \item{description}{ ... }
  \item{K.A.}{...}
  \item{draft}{ ... }
  \item{revised}{ ... }
  \item{reserved}{ ... }
  \item{approved}{ ... }
  \item{executed}{ ... }
}
}
\keyword{datasets}

