% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_data.R
\name{city_data}
\alias{city_data}
\title{city data example}
\format{A data frame with the previous characteristics as columns}
\description{
This dataset is an example data frame of the budget phase data

\itemize{
  \item Administrative_Unit
  \item Approved
  \item Draft
  \item Executed
  \item Revised
  }
}
