% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl.analysis.r
\name{cl.analysis}
\alias{cl.analysis}
\title{Cluster OBEU}
\usage{
cl.analysis(cluster.data,cluster.method=NULL,cluster.number=NULL,distance="euclidean")
}
\arguments{
\item{cluster.data}{The input data}

\item{cluster.method}{The clustering algorithm}

\item{cluster.number}{The number of clusters}

\item{distance}{The distance metric}
}
\value{
This function returns the graphical and evaluation paremeters of the selected algorithm.
}
\description{
Clustering Methods for OBEU datasets.
}
\details{
...
}
\examples{
cl.analysis(Budget_Thessaloniki_2015_Expenditure, cl_feature=NULL, measured.dimension="variable", cl.aggregate="sum",
               cluster.method="pam", cluster.number=3, distance="euclidean")

}
\author{
Kleanthis Koupidis, Jaroslav Kuchar
}
\seealso{
\code{\link{cl.feature}}
}

