% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl.feature.R
\name{cl.feature}
\alias{cl.feature}
\title{Clustering feature}
\usage{
cl.feature(data, feature=NULL, measured="variable",aggregate="sum")
}
\arguments{
\item{data}{The input data}

\item{feature}{The clustering feature}

\item{measured}{The measured dimension of dataset}

\item{aggregate}{The function to aggregate}
}
\value{
This function returns the dataset for cluster analysis adapted to the desired feature.
}
\description{
Select clustering characteristic of OBEU datasets.
}
\details{
This function adapts the dataset according to the selected dimension of the dataset 
and the aggregation function.
}
\examples{
cl.feature(Budget_Thessaloniki_2015_Revenue, feature=NULL, measured="variable", aggregate="sum")

}
\author{
Kleanthis Koupidis
}
\seealso{
\code{\link{cl.analysis}}
}

