% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl.feature.R
\name{cl.features}
\alias{cl.features}
\title{Clustering features}
\usage{
cl.features(data, features = NULL, amounts = NULL, aggregate = "sum", tojson = FALSE )
}
\arguments{
\item{data}{The input data}

\item{features}{The clustering features}

\item{amounts}{The amount measures of the dataset}

\item{aggregate}{The function to aggregate}

\item{tojson}{If TRUE the results are returned in json format, default returns a list}
}
\value{
This function returns the dataset for cluster analysis adapted to the desired features.
}
\description{
Select clustering characteristic to form the clustering data
}
\details{
This function adapts the dataset according to the selected dimension of the dataset 
and the aggregation function.
}
\examples{
cl.features(city_data, features = 'Administrative_Unit')

# works also for other datasets
cl.features(iris, features = 'Species')
}
\seealso{
\code{\link{cl.analysis}}
}
\author{
Kleanthis Koupidis
}
