% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl.plot.R
\name{cl.plot}
\alias{cl.plot}
\title{Plot of clustering model}
\usage{
cl.plot(clustering.model, parameters = list())
}
\arguments{
\item{clustering.model}{...}

\item{parameters}{...}
}
\value{
...
}
\description{
...
}
\examples{
data("iris")
inputs.data <- scale(iris[,1:4])
inputs.clustering <- cl.analysis(inputs.data, cluster.method="kmeans", cluster.number=3)
cl.plot(inputs.clustering, parameters = list(convex.hulls=TRUE))
}
\author{
Jaroslav Kuchar <https://github.com/jaroslav-kuchar>
}

