% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl.plot.R
\name{convex.hulls}
\alias{convex.hulls}
\title{Convex hull points}
\usage{
convex.hulls(clustering.model, data.pca)
}
\arguments{
\item{clustering.model}{Object returned by the \code{\link{cl.analysis}} function.}

\item{data.pca}{data as result of the \code{stats::prcomp(clustering.model$data, scale. = T, center = T)}.}
}
\value{
List of vectors with points for each convex hull.
}
\description{
Computes points to plot a convex hull for each cluster of the clustering model
}
