% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nums.R
\name{nums}
\alias{nums}
\title{Select the numeric columns of a given dataset}
\usage{
nums(data)
}
\arguments{
\item{data}{The input data frame, matrix}
}
\value{
This function returns a data frame with the numeric columns of the input dataset.
}
\description{
Extract and return a data frame with the columns that include only numeric values
}
\examples{
nums(city_data)
}
\author{
Kleanthis Koupidis
}
