% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.clustering.model.R
\name{plot.clustering.model}
\alias{plot.clustering.model}
\title{Plot of clustering model}
\usage{
plot.clustering.model(clustering.model, parameters = list())
}
\arguments{
\item{clustering.model}{...}

\item{parameters}{...}
}
\value{
...
}
\description{
...
}
\examples{
data("iris")
inputs.data <- scale(iris[,1:4])
inputs.clustering <- cl.analysis(inputs.data,"kmeans",3)
plot.clustering.model(inputs.clustering, parameters = list(convex.hulls=TRUE))
}
\author{
Jaroslav Kuchar <https://github.com/jaroslav-kuchar>
}

