% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposed.meth.nb.clust.R
\name{proposed.meth.nb.clust}
\alias{proposed.meth.nb.clust}
\title{Extract the proposed clustering method and the number of clusters from clvalid method}
\usage{
proposed.meth.nb.clust(clv)
}
\arguments{
\item{clv}{A clValid object}
}
\value{
...
}
\description{
...
}
\details{
...
}
\author{
Kleanthis Koupidis
}
\seealso{
proposed.methclust, proposed.nbclust
}

