% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposed.methclust.R
\name{proposed.methclust}
\alias{proposed.methclust}
\title{Extract the proposed clustering method from clvalid method}
\usage{
proposed.methclust(clv)
}
\arguments{
\item{clv}{A clValid object}
}
\value{
...
}
\description{
...
}
\details{
...
}
\author{
Kleanthis Koupidis
}
\seealso{
proposed.meth.nb.clust, proposed.nbclust
}

