% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposed.nbclust.R
\name{proposed.nbclust}
\alias{proposed.nbclust}
\title{Extract the proposed number of clusters from clvalid method}
\usage{
proposed.nbclust(clv)
}
\arguments{
\item{clv}{A clValid object}
}
\value{
...
}
\description{
...
}
\details{
...
}
\seealso{
proposed.meth.nb.clust, proposed.methclust
}
\author{
Kleanthis Koupidis
}
