% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample city data.R
\name{sample_city_data}
\alias{sample_city_data}
\title{Sample city data}
\format{A data frame with the previous characteristics as columns}
\description{
This dataset is an example data frame of the budget phase data

\itemize{
  \item Administrative_Unit
  \item Approved
  \item Draft
  \item Executed
  \item Revised
  }
}
