% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.obeu.R
\name{ts.obeu}
\alias{ts.obeu}
\title{Time series analysis results for OBEU Time series}
\usage{
ts.obeu(tsdata, prediction_steps)
}
\arguments{
\item{tsdata}{The input univariate time series data}

\item{prediction_steps}{The number of prediction steps}
}
\value{
A json string with the resulted parameters.
}
\description{
Univariate time series analysis for short and long time series different models implemented.
}
\details{
Works perfect but still needs more.
This function automatically selects the appropriate arima model that fits the input data.
The model selection depends on the results of some diagnostic tests (acf,pacf,pp adf and kpss).
}
\examples{
#Not YET an OBeu Example
ts.obeu(Athens_draft_ts,1)
ts.obeu(Athens_revised_ts,2)
ts.obeu(Athens_reserved_ts,3)
ts.obeu(Athens_approved_ts,4)
ts.obeu(Athens_executed_ts,5)

}
\author{
Kleanthis Koupidis
}
\references{
add
}
\seealso{
add
}

