% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl.analysis.r
\name{cl.analysis}
\alias{cl.analysis}
\title{Cluster OBEU}
\usage{
cl.analysis(cl.data, cl_feature=NULL, measured.dim="variable", cl.aggregate="sum",
cl.meth=NULL, clust.numb=NULL, dist="euclidean")
}
\arguments{
\item{cl.data}{The input data}

\item{cl_feature}{...}

\item{measured.dim}{...}

\item{cl.aggregate}{...}

\item{cl.meth}{The clustering method algorithm}

\item{clust.numb}{The number of clusters}

\item{dist}{The dist metric}
}
\value{
This function returns the graphical and evaluation paremeters of the selected algorithm.
}
\description{
Clustering Analysis for OBEU datasets.
}
\details{
...
}
\author{
Kleanthis Koupidis, Jaroslav Kuchar
}
\seealso{
\code{\link{cl.feature}}
}

