% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl.feature.R
\name{cl.features}
\alias{cl.features}
\title{Clustering features}
\usage{
cl.features(data, features=NULL, amounts=NULL, aggregate="sum")
}
\arguments{
\item{data}{The input data}

\item{features}{The clustering features}

\item{amounts}{The amount measures of the dataset}

\item{aggregate}{The function to aggregate}
}
\value{
This function returns the dataset for cluster analysis adapted to the desired features.
}
\description{
Select clustering characteristic of OBEU datasets.
}
\details{
This function adapts the dataset according to the selected dimension of the dataset 
and the aggregation function.
}
\seealso{
\code{\link{cl.analysis}}
}
\author{
Kleanthis Koupidis
}
