% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl.summary.R
\name{cl.summary}
\alias{cl.summary}
\title{Extract the proposed clustering method and the number of clusters from clvalid method}
\usage{
cl.summary(clv)
}
\arguments{
\item{clv}{A clValid object}
}
\value{
A value that indicates the proposed method and number of clusters.
}
\description{
Extract the most frequent
}
\details{
This function returns the proposed method or number of clusters or both according to the majority clustering indices of a clValid process
}
\author{
Kleanthis Koupidis
}
