% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_spending.cl.R
\name{open_spending.cl}
\alias{open_spending.cl}
\title{Read and Calculate the Basic Information for Cluster Analysis Tasks from Open Spending API}
\usage{
open_spending.cl(json_data, dimensions=NULL, amounts=NULL, measured.dim=NULL,
                        cl.feature=NULL, cl.measured.dim="variable", cl.aggregate="sum",
                        cl.method=NULL, cl.num=NULL, cl.dist="euclidean")
}
\arguments{
\item{json_data}{The json string, URL or file from Open Spending API}

\item{dimensions}{The dimensions of the input data}

\item{amounts}{The measures of the input data}

\item{measured.dim}{The dimensions to which correspond amount/numeric variables}

\item{cl.feature}{...}

\item{cl.measured.dim}{...}

\item{cl.aggregate}{...}

\item{cl.method}{...}

\item{cl.num}{...}

\item{cl.dist}{...}
}
\value{
A json string with the resulted parameters of the \code{\link{cl.analysis}} function.
}
\description{
Extract and analyze the input data provided from Open Spending API, using the \code{\link{cl.analysis}} function.
}
\details{
This function is used to read data in json format from Open Spending API, in order to implement 
some cluster analysis tasks through \code{\link{cl.analysis}} function.
}
\author{
Kleanthis Koupidis
}
\seealso{
\code{\link{cl.analysis}}
}

