% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposed.meth.nb.clust.R
\name{proposed.meth.nb.clust}
\alias{proposed.meth.nb.clust}
\title{Extract the proposed clustering method and the number of clusters from clvalid method}
\usage{
proposed.meth.nb.clust(clv)
}
\arguments{
\item{clv}{A clValid object}
}
\value{
A value that indicates the proposed method and number of clusters.
}
\description{
...
}
\details{
This function returns the proposed method or number of clusters or both according to the majority clustering indices of a clValid process
}
\author{
Kleanthis Koupidis
}
